backend/requirements.txt
fastapi==0.115.6
uvicorn[standard]==0.34.0
python-multipart==0.0.20
python-docx==1.1.2
chromadb==0.6.3
sentence-transformers==3.4.1
openai==1.59.9
httpx==0.28.1
pydantic==2.10.5
pydantic-settings==2.7.1
